Revert "CB-25773 cleaning up sh files in the CB repo"
This reverts commit e9f95c60f84ed8ddfa6d1de6765e28b40b003543.

+++ integration-test/scripts/docker-compose-after-config-change.sh
@@ -56,46 +56,49 @@ else
     PUBLIC_IP=127.0.0.1
 fi
 
-date
-while read line; do
-export $line
-done < integrationtest.properties
+if [[ "$CIRCLECI" && "$CIRCLECI" == "true" ]]; then
 
-date
-echo -e "\n\033[1;96m--- Env variables started with INTEGRATIONTEST :\033[0m\n"
-env | grep -i INTEGRATIONTEST
+    date
+    while read line; do
+        export $line
+    done < integrationtest.properties
 
-date
-echo -e "\n\033[1;96m--- Tests to run:\033[0m\n"
-echo $INTEGRATIONTEST_SUITEFILES
+    date
+    echo -e "\n\033[1;96m--- Env variables started with INTEGRATIONTEST :\033[0m\n"
+    env | grep -i INTEGRATIONTEST
 
-date
-echo -e "\n\033[1;96m--- Start testing... (it may take few minutes to finish.)\033[0m\n"
-rm -rf test-output
+    date
+    echo -e "\n\033[1;96m--- Tests to run:\033[0m\n"
+    echo $INTEGRATIONTEST_SUITEFILES
 
-export DOCKER_CLIENT_TIMEOUT=120
-export COMPOSE_HTTP_TIMEOUT=120
+    date
+    echo -e "\n\033[1;96m--- Start testing... (it may take few minutes to finish.)\033[0m\n"
+    rm -rf test-output
 
-set -o pipefail ; docker compose --compatibility up --remove-orphans --exit-code-from test test | tee test_2.out
-echo -e "\n\033[1;96m--- Test finished\033[0m\n"
+    export DOCKER_CLIENT_TIMEOUT=120
+    export COMPOSE_HTTP_TIMEOUT=120
 
-echo -e "\n\033[1;96m--- Collect docker stats:\033[0m\n"
-if [[ -z "${INTEGRATIONTEST_YARN_QUEUE}" ]] && [[ "$AWS" != true ]]; then
-  sudo mkdir -p ./test-output
-  sudo chmod -R a+rwx ./test-output
-  sudo chmod -R a+rwx ./integcb/logs
-  mkdir ./test-output/docker_stats
-  docker stats --no-stream --format "{{ .NetIO }}" cbreak_commondb_1 > ./test-output/docker_stats/pg_stat_network_io.result;
+    set -o pipefail ; docker compose --compatibility up --remove-orphans --exit-code-from test test | tee test_2.out
+    echo -e "\n\033[1;96m--- Test finished\033[0m\n"
 
-  docker stats --no-stream --format "table {{ .Name }}\t{{ .Container }}\t{{ .MemUsage }}\t{{ .MemPerc }}\t{{ .CPUPerc }}\t{{ .NetIO }}\t{{ .BlockIO }}" > ./test-output/docker_stats/docker_stat.html
-  docker exec cbreak_commondb_1 psql -U postgres --pset=pager=off -d cbdb -c "CREATE EXTENSION IF NOT EXISTS pg_stat_statements;";
-  docker exec cbreak_commondb_1 psql -U postgres --pset=pager=off -d cbdb -c "select * from pg_stat_statements;" --html > ./test-output/docker_stats/query_stat.html
+    echo -e "\n\033[1;96m--- Collect docker stats:\033[0m\n"
+    if [[ -z "${INTEGRATIONTEST_YARN_QUEUE}" ]] && [[ "$AWS" != true ]]; then
+        sudo mkdir -p ./test-output
+        sudo chmod -R a+rwx ./test-output
+        sudo chmod -R a+rwx ./integcb/logs
+        mkdir ./test-output/docker_stats
+        docker stats --no-stream --format "{{ .NetIO }}" cbreak_commondb_1 > ./test-output/docker_stats/pg_stat_network_io.result;
 
-  cp ./src/main/resources/pg_stats/pg_query_stat_template.html ./test-output/docker_stats/pg_query_stat_template.html
+        docker stats --no-stream --format "table {{ .Name }}\t{{ .Container }}\t{{ .MemUsage }}\t{{ .MemPerc }}\t{{ .CPUPerc }}\t{{ .NetIO }}\t{{ .BlockIO }}" > ./test-output/docker_stats/docker_stat.html
+        docker exec cbreak_commondb_1 psql -U postgres --pset=pager=off -d cbdb -c "CREATE EXTENSION IF NOT EXISTS pg_stat_statements;";
+        docker exec cbreak_commondb_1 psql -U postgres --pset=pager=off -d cbdb -c "select * from pg_stat_statements;" --html > ./test-output/docker_stats/query_stat.html
 
-  #FIXME, might be better not to use in place sed
-  sed -i '/<!-- CB_PG_STAT -->/r ./test-output/docker_stats/query_stat.html' ./test-output/docker_stats/pg_query_stat_template.html
-  sed -i '/<!-- DOCKER_STAT_RESULT -->/r ./test-output/docker_stats/docker_stat.html' ./test-output/docker_stats/pg_query_stat_template.html
+        cp ./src/main/resources/pg_stats/pg_query_stat_template.html ./test-output/docker_stats/pg_query_stat_template.html
 
-  mv ./test-output/docker_stats/pg_query_stat_template.html ./test-output/docker_stats/query_stat.html
+        #FIXME, might be better not to use in place sed
+        sed -i '/<!-- CB_PG_STAT -->/r ./test-output/docker_stats/query_stat.html' ./test-output/docker_stats/pg_query_stat_template.html
+        sed -i '/<!-- DOCKER_STAT_RESULT -->/r ./test-output/docker_stats/docker_stat.html' ./test-output/docker_stats/pg_query_stat_template.html
+
+        mv ./test-output/docker_stats/pg_query_stat_template.html ./test-output/docker_stats/query_stat.html
+    fi
 fi
